+++ changelog.md
@@ -20,6 +20,7 @@ Uiua is not yet stable.
 - LSP improvements
   - Add hover information on binding references
   - Add signatures to binding hover information
+  - Add same-file binding rename support
 - Add the `uiua build` command, which emits a `.uasm` bytecode file
 - `uiua run` can now run a `.uasm` bytecode file
 - Bug and crash fixes